Maintain `int32` as a legitimate type for `Policy`
There are places in the code where we do need to do floating point operations on policy values (see around line 408 of `StationaryDist_FHorz_Case1`, but this can be handled in the moment and does not foreclose on maintaining `int32` Policies generally.

+++ StationaryDist/FHorz/StationaryDist_FHorz_Case1.m
@@ -405,7 +405,7 @@ else
         % Policy_aprime(:,:,1,:) lower grid point for a1 is unchanged
         Policy_aprime(:,:,2,:)=Policy_aprime(:,:,2,:)+1; % add one to a1, to get upper grid point
 
-        aprimeProbs_upper=reshape(shiftdim((Policy(end-1,:,:,:)-1)/(simoptions.ngridinterp+1),1),[N_a,N_ze,1,N_j]); % probability of upper grid point (from L2 index; end-1 because end is now L2flag)
+        aprimeProbs_upper=reshape(shiftdim(double(Policy(end-1,:,:,:)-1)/(simoptions.ngridinterp+1),1),[N_a,N_ze,1,N_j]); % probability of upper grid point (from L2 index; end-1 because end is now L2flag)
         PolicyProbs(:,:,1,:)=PolicyProbs(:,:,1,:).*(1-aprimeProbs_upper); % lower a1
         PolicyProbs(:,:,2,:)=PolicyProbs(:,:,2,:).*aprimeProbs_upper; % upper a1
